Updated react-datepicker, input value can now be typed freely
