+++ package.json
@@ -3,7 +3,7 @@
     "version": "0.5.0",
     "private": true,
     "dependencies": {
-        "@types/react-datepicker": "^2.9.3",
+        "@types/react-datepicker": "^2.9.5",
         "apollo-boost": "^0.1.10",
         "apollo-cache-inmemory": "^1.2.9",
         "apollo-client": "^2.4.1",
@@ -48,7 +48,7 @@
         "raf": "3.4.0",
         "react": "^16.4.1",
         "react-apollo": "^2.1.9",
-        "react-datepicker": "^2.9.6",
+        "react-datepicker": "^2.10.1",
         "react-dom": "^16.4.1",
         "react-icons": "3.0.5",
         "react-leaflet": "^2.0.1",
